Buncha work today :) - Added blocks - Added tags to those blocks - Added Tag data gen, both for blocks and items
I reckon tomorrow I'll probably rip the band-aid off and refactor this entire workspace + the repo to "The Noxerna" because that's what this test mod will probably end up being for later on, and I'd rather make a big change like that before it bites me in the ass. Other than that, will still be slowly chipping away at the things I need to do.